+++ CHANGES.md
@@ -3,3 +3,4 @@
 ## v0.1.1
 
 - Added option `connectTimeout` for TCP connection provider.
+- Fixed the method `close` for TCP connection provider.